requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
paho-mqtt